--- apps/web/app/select/head.tsx
@@ -1,7 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>A or B 뭐가 나을까요?</title>
-    </>
-  );
-}
